requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anchore_syft==1.18.1
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
asttokens @ file:///croot/asttokens_1743630435401/work
asyncache==0.3.1
attrs==25.4.0
azure-core==1.36.0
azure-storage-blob==12.19.1
bcrypt==4.1.2
boto3==1.34.56
botocore==1.34.162
bracex==2.6
cachetools==5.5.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
contourpy==1.3.2
croniter==6.0.0
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.17
decorator @ file:///home/task_175734035162575/conda-bld/decorator_1757341232127/work
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
dynaconf==3.2.6
email-validator==2.3.0
exceptiongroup==1.3.0
executing @ file:///home/task_175706124557094/conda-bld/executing_1757061259817/work
fastapi==0.115.7
fastuuid==0.14.0
filelock==3.20.0
fonttools==4.60.1
forbiddenfruit==0.1.4
frozenlist==1.8.0
fsspec==2025.10.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
importlib_metadata==8.5.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.26.0
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
ipywidgets==8.1.2
isodate==0.7.2
itables==1.7.1
jedi @ file:///croot/jedi_1733987392413/work
Jinja2==3.1.5
jiter==0.11.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1751991368470/work
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kr8s==0.13.5
linkify-it-py==2.0.3
litellm==1.79.1
loguru==0.7.3
Markdown==3.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.0
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.5
nh3==0.2.17
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
openai==2.7.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-dbapi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlite3==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging @ file:///home/task_176104885106445/conda-bld/packaging_1761049078006/work
pandas==2.2.2
parso==0.8.5
pathspec==0.12.1
pexpect @ file:///croot/pexpect_1750072593878/work
pid==3.0.4
pillow==12.0.0
platformdirs @ file:///home/task_176235638439362/conda-bld/platformdirs_1762356487841/work
pluggy==1.6.0
prompt_toolkit @ file:///home/task_176174483353592/conda-bld/prompt-toolkit_1761744845973/work
propcache==0.4.1
protobuf==5.29.5
psutil==6.1.1
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.7
psycopg2-binary==2.9.9
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/task_175706703828594/conda-bld/pure_eval_1757067053474/work
py_fast_rsync==0.1.0
pyarrow==17.0.0
pycapnp==2.0.0
pycparser==2.23
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.5
pytest==9.0.0
python-box==7.3.2
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.2.1
python-jsonpath==2.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.1.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.3
RestrictedPython==7.0
rich==13.9.4
rpds-py==0.28.0
s3transfer==0.10.4
shellingham==1.5.4
sherlock==0.4.1
six @ file:///croot/six_1744271502820/work
sniffio==1.3.1
SQLAlchemy==2.0.32
stack_data @ file:///home/task_175706683579810/conda-bld/stack_data_1757067036897/work
starlette==0.45.3
syft-core==0.1.0
syft-event==0.1.0
syft-rpc==0.1.0
syftbox==0.3.5
sympy==1.13.1
tenacity==8.3.0
textual==6.2.1
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.0.1
tomli_w==1.0.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
triton==3.2.0
typeguard==4.1.5
typer==0.15.1
typer-slim==0.20.0
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.34.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcmatch==10.0
wcwidth==0.2.14
websockets==15.0.1
widgetsnbextension==4.0.15
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
